REFACTOR: fold the store into motion_sequence, name the two directions symmetrically
The motion sequence and its .rrd encoding are one thing described twice, so rrd_io is
merged into motion_sequence: MotionSequence (in memory) and MotionStore (on disk) now sit
together, and ARRAY_FIELDS is the single place the field list is written down, used by
copy(), the store and the migrate command alike.

MotionSequence stays plain numpy rather than becoming a live Rerun store. Around 45
producer sites fill it element by element (`motion.body_positions[f, i] = ...`); Arrow
chunks are immutable and column-oriented, so backing it with one would make every
converter slower and clumsier to write, and buy nothing — the round trip is already exact.

Naming now says which side of the hub a function is on: motion_from_X produces a motion,
X_from_motion consumes one.

    build_body_motion_data  -> motion_from_armature      (animation -> robot, into the hub)
    motion_from_policy_log  -> motion_from_robot_log     (robot -> animation, into the hub)
    load_motion_to_armature -> armature_from_motion      (robot -> animation, out of the hub)
    read_motion_mcap        -> read_robot_log
    bake_retarget           -> retarget_armature         (mirror of MotionRetargeting)

Each converter's docstring states its direction and names its opposite number, and the
README carries a step-by-step table of the two pipelines side by side.

Also drops a wm.redraw_timer call from motion_from_armature. It was there to force pose
matrices to refresh, but it cannot run in background Blender, which is why the export
scripts had always needed the GUI. frame_set plus view_layer.update() is enough on Blender
5.x: re-exporting zamuza headless gives 2000 distinct frames out of 2001, with rotations
varying, so nothing was stale. MotionRetargeting's Args block also still described three
parameters that no longer exist.

The retired .npz files are deleted. Nothing writes that format now; the only code that
reads it is `mikumotion migrate`, kept because the Drive mirror still holds .npz motions.
